+++ android/app/src/main/java/com/example/water_harvester/Classes/GraphBuilder.java
@@ -16,7 +16,7 @@ import lecho.lib.hellocharts.model.SliceValue;
 import lecho.lib.hellocharts.view.PieChartView;
 
 public class GraphBuilder {
-    private static PieChartData BuildPieChartData(int value, String label, String centerText1, String centerText2, int fontSize, int color){
+    private static PieChartData BuildPieChartData(long value, String label, String centerText1, String centerText2, int fontSize, int color){
         List<SliceValue> pieData = new ArrayList<>();
             pieData.add(new SliceValue(value).setColor(color).setLabel(label));
             pieData.add(new SliceValue(100-value).setLabel(""));
@@ -30,23 +30,23 @@ public class GraphBuilder {
                 .setCenterText2FontSize(fontSize);
     }
 
-    public static PieChartData BuildAmbientTempPieChartData(int value, int color){
+    public static PieChartData BuildAmbientTempPieChartData(long value, int color){
         return BuildPieChartData(value * 2, value + "°C", "Ambient", "temperature", 10, color);
     }
 
-    public static PieChartData BuildAmbientHumPieChartData(int value, int color){
+    public static PieChartData BuildAmbientHumPieChartData(long value, int color){
         return BuildPieChartData(value, value + "%", "Ambient", "humidity", 10, color);
     }
 
-    public static PieChartData BuildInsideTempPieChartData(int value, int color){
+    public static PieChartData BuildInsideTempPieChartData(long value, int color){
         return BuildPieChartData(value * 2, value + "°C", "Inside", "temperature", 10, color);
     }
 
-    public static PieChartData BuildInsideHumPieChartData(int value, int color){
+    public static PieChartData BuildInsideHumPieChartData(long value, int color){
         return BuildPieChartData(value, value + "%", "Inside", "humidity", 10, color);
     }
 
-    public static PieChartData BuildFanRPMPieChartData(int value, int color){
+    public static PieChartData BuildFanRPMPieChartData(long value, int color){
         return BuildPieChartData(value, value * 25 + "rpm", "Fan", "RPM", 10, color);
     }
 }
